+++ DesktopApplicationTemplate.UI/ViewModels/ScpEditServiceViewModel.cs
@@ -1,5 +1,4 @@
 using System;
-using System.Windows.Input;
 using DesktopApplicationTemplate.Core.Services;
 using DesktopApplicationTemplate.UI.Services;
 
@@ -8,49 +7,104 @@ namespace DesktopApplicationTemplate.UI.ViewModels;
 /// <summary>
 /// View model for editing an existing SCP service configuration.
 /// </summary>
-public class ScpEditServiceViewModel : ScpCreateServiceViewModel
+public class ScpEditServiceViewModel : ServiceEditViewModelBase<ScpServiceOptions>
 {
+    private readonly ScpServiceOptions _options;
+    private string _serviceName;
+    private string _host;
+    private string _port;
+    private string _username;
+    private string _password;
+
     /// <summary>
     /// Initializes a new instance of the <see cref="ScpEditServiceViewModel"/> class.
     /// </summary>
-    public ScpEditServiceViewModel(ILoggingService? logger = null)
+    public ScpEditServiceViewModel(string serviceName, ScpServiceOptions options, ILoggingService? logger = null)
         : base(logger)
     {
+        _options = options ?? throw new ArgumentNullException(nameof(options));
+        _serviceName = serviceName ?? throw new ArgumentNullException(nameof(serviceName));
+        _host = options.Host;
+        _port = options.Port.ToString();
+        _username = options.Username;
+        _password = options.Password;
     }
-
-    /// <summary>
-    /// Loads existing options into the view model for editing.
-    /// </summary>
-    /// <param name="serviceName">Existing service name.</param>
-    /// <param name="options">Current SCP options to edit.</param>
-    public void Load(string serviceName, ScpServiceOptions options)
-    {
-        if (options is null) throw new ArgumentNullException(nameof(options));
-
-        ServiceName = serviceName;
-        Host = options.Host;
-        Port = options.Port.ToString();
-        Username = options.Username;
-        Password = options.Password;
-        Options.Host = options.Host;
-        Options.Port = options.Port;
-        Options.Username = options.Username;
-        Options.Password = options.Password;
-        Options.LocalPath = options.LocalPath;
-        Options.RemotePath = options.RemotePath;
-    }
-
-    /// <summary>
-    /// Command for saving the updated configuration.
-    /// </summary>
-    public ICommand SaveCommand => CreateCommand;
 
     /// <summary>
     /// Raised when the configuration is saved.
     /// </summary>
-    public event Action<string, ScpServiceOptions>? ServiceUpdated
+    public event Action<string, ScpServiceOptions>? ServiceUpdated;
+
+    /// <summary>
+    /// Raised when editing is cancelled.
+    /// </summary>
+    public event Action? Cancelled;
+
+    /// <summary>
+    /// Raised when advanced configuration is requested.
+    /// </summary>
+    public event Action<ScpServiceOptions>? AdvancedConfigRequested;
+
+    /// <summary>
+    /// Name of the service.
+    /// </summary>
+    public string ServiceName
     {
-        add => ServiceCreated += value;
-        remove => ServiceCreated -= value;
+        get => _serviceName;
+        set { _serviceName = value; OnPropertyChanged(); }
     }
+
+    /// <summary>
+    /// SCP host name.
+    /// </summary>
+    public string Host
+    {
+        get => _host;
+        set { _host = value; OnPropertyChanged(); }
+    }
+
+    /// <summary>
+    /// SCP port number as string.
+    /// </summary>
+    public string Port
+    {
+        get => _port;
+        set { _port = value; OnPropertyChanged(); }
+    }
+
+    /// <summary>
+    /// Username for authentication.
+    /// </summary>
+    public string Username
+    {
+        get => _username;
+        set { _username = value; OnPropertyChanged(); }
+    }
+
+    /// <summary>
+    /// Password for authentication.
+    /// </summary>
+    public string Password
+    {
+        get => _password;
+        set { _password = value; OnPropertyChanged(); }
+    }
+
+    /// <inheritdoc />
+    protected override void OnSave()
+    {
+        _options.Host = Host;
+        if (int.TryParse(Port, out var port))
+            _options.Port = port;
+        _options.Username = Username;
+        _options.Password = Password;
+        ServiceUpdated?.Invoke(ServiceName, _options);
+    }
+
+    /// <inheritdoc />
+    protected override void OnCancel() => Cancelled?.Invoke();
+
+    /// <inheritdoc />
+    protected override void OnAdvancedConfig() => AdvancedConfigRequested?.Invoke(_options);
 }
+